requirements.txt
django
django-rest-framework
djangorestframework-simplejwt
django-allauth
python-dotenv
